+++ inst/create_hex.R
@@ -1,0 +1,14 @@
+
+
+sticker("inst/figures/discoball_long.png",
+        package = "iwillsurvive",
+        p_color = "white",
+        h_color = "white",
+        p_size = 6, s_x = 1.03, s_y = 1.1,
+        h_fill = gray(.5),
+        s_width = .75,
+        url = "https://github.com/ndphillips/iwillsurvive",
+        u_color = "white",
+        u_size = 1.25,
+        asp = .85,
+        filename = "inst/figures/iwillsurvive_hex.png")