+++ NEWS.md
@@ -1,3 +1,12 @@
+# iwillsurvive dev
+
+* Renamed fit_survival -> iwillsurvive()
+* Allow users to pass an existing survfit object to iwillsurvive()
+* Added new hex sticker usin the hexSticker package
+
+# iwillsurvive 0.1.1
+
+
 # iwillsurvive 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.